notification_microservice/requirements/base.txt
Django==2.0.8
djangorestframework==3.8.2
Pillow==5.2.0
requests>=2.20.0
psycopg2-binary==2.7.5
python-decouple==3.1
exponent_server_sdk==0.2.0
